requirements.txt
# Core dependencies
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0

# OpenAI
openai>=1.12.0

# Async support
asyncio>=3.4.3
aiofiles>=23.2.1

# File handling
PyPDF2>=3.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=6.0.0
openpyxl>=3.1.5

# Type hints support
typing-extensions>=4.8.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0
pytest-mock>=3.14.0
coverage>=7.3.0

#Google
google-api-python-client>=2.161.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1

#utils
langdetect>=1.0.9
deep-translator>=1.11.4